+++ src/draw.cpp
@@ -251,7 +251,7 @@ void Draw_Pic(int x, int y, qpic_t* pic)
     unsigned short* pusdest;
     int v, u;
 
-    if ((x < 0) || (x + pic->width > vid.width) || (y < 0) || (y + pic->height > vid.height)) {
+    if ((x < 0) || (unsigned)(x + pic->width) > vid.width || (y < 0) || (unsigned)(y + pic->height) > vid.height) {
         Sys_Error("Draw_Pic: bad coordinates");
     }
 
@@ -519,7 +519,7 @@ void Draw_ConsoleBackground(int lines)
             } else {
                 f = 0;
                 fstep = 320 * 0x10000 / vid.conwidth;
-                for (x = 0; x < vid.conwidth; x += 4) {
+                for (x = 0; x < (int)vid.conwidth; x += 4) {
                     dest[x] = src[f >> 16];
                     f += fstep;
                     dest[x + 1] = src[f >> 16];
@@ -541,7 +541,7 @@ void Draw_ConsoleBackground(int lines)
             src = conback->data + v * 320;
             f = 0;
             fstep = 320 * 0x10000 / vid.conwidth;
-            for (x = 0; x < vid.conwidth; x += 4) {
+            for (x = 0; x < (int)vid.conwidth; x += 4) {
                 pusdest[x] = d_8to16table[src[f >> 16]];
                 f += fstep;
                 pusdest[x + 1] = d_8to16table[src[f >> 16]];
@@ -762,13 +762,13 @@ void Draw_FadeScreen(void)
     S_ExtraUpdate();
     VID_LockBuffer();
 
-    for (y = 0; y < vid.height; y++) {
+    for (y = 0; y < (int)vid.height; y++) {
         int t;
 
         pbuf = (byte*)(vid.buffer + vid.rowbytes * y);
         t = (y & 1) << 1;
 
-        for (x = 0; x < vid.width; x++) {
+        for (x = 0; x < (int)vid.width; x++) {
             if ((x & 3) != t) {
                 pbuf[x] = 0;
             }
